p4i integration Integrate 259604 from 4.3 to main: Fixes:    intel compiler warning Details:
       unsigned all the way, baby

[git-p4: depot-paths = "//depot/qt/main/": change = 259608]

+++ src/gui/image/qimage.cpp
@@ -3312,7 +3312,7 @@ void QImage::setPixel(int x, int y, uint index_or_rgb)
     case Format_RGB32:
         //make sure alpha is 255, we depend on it in qdrawhelper for cases
         // when image is set as a texture pattern on a qbrush
-        ((uint *)s)[x] = (255 << 24) | index_or_rgb;
+        ((uint *)s)[x] = uint(255 << 24) | index_or_rgb;
         break;
     case Format_ARGB32:
     case Format_ARGB32_Premultiplied:
